refactor: split statistics endpoint to support versioning
